Refactor API image parameters methods and split in 3 modules (input/output/processing)
