+++ pims/api/thumb.py
@@ -18,10 +18,9 @@ from starlette.responses import Response
 
 from pims.api.exceptions import check_representation_existence
 from pims.api.utils.header import ImageRequestHeaders, add_image_size_limit_header
-from pims.api.utils.image_parameter import (
-    check_array_size, check_reduction_validity, get_channel_indexes, get_thumb_output_dimensions,
-    get_timepoint_indexes, get_zslice_indexes,
-    parse_colormap_ids, parse_filter_ids, parse_intensity_bounds, safeguard_output_dimensions
+from pims.api.utils.input_parameter import (
+    check_reduction_validity, get_channel_indexes, get_timepoint_indexes,
+    get_zslice_indexes
 )
 from pims.api.utils.mimetype import (
     OutputExtension, VISUALISATION_MIMETYPES,
@@ -31,14 +30,22 @@ from pims.api.utils.models import (
     ChannelReduction, ImageOpsDisplayQueryParams, ImageOutDisplayQueryParams,
     PlaneSelectionQueryParams, ThumbnailRequest
 )
+from pims.api.utils.output_parameter import (
+    get_thumb_output_dimensions,
+    safeguard_output_dimensions
+)
 from pims.api.utils.parameter import imagepath_parameter
+from pims.api.utils.processing_parameter import (
+    parse_colormap_ids, parse_filter_ids,
+    parse_intensity_bounds
+)
 from pims.cache import cache_image_response
 from pims.config import Settings, get_settings
 from pims.files.file import Path
 from pims.filters import FILTERS
 from pims.processing.colormaps import ALL_COLORMAPS
 from pims.processing.image_response import ThumbnailResponse
-from pims.utils.iterables import ensure_list
+from pims.utils.iterables import check_array_size, ensure_list
 
 router = APIRouter()
 api_tags = ['Thumbnails']
